commited

+++ src/infra/db/mongodb/repositories/mongo.repository.ts
@@ -16,6 +16,8 @@ const cryptService = new BcryptService();
 import userModel from '../drivers/mongoose/models/User';
 import groupModel from '../drivers/mongoose/models/Group';
 import eventModel from '../drivers/mongoose/models/Event';
+import groupLeadersModel from '../drivers/mongoose/models/Group_Leader';
+import groupSubscriberModel from '../drivers/mongoose/models/Group_Subscribers';
 
 export default class MongoRepository implements IUserRepository {
     async save(user: User): Promise<boolean> {
@@ -41,10 +43,21 @@ export default class MongoRepository implements IUserRepository {
 
     async saveGroup(group: Group): Promise<boolean> {
         await groupModel.create(group);
+        const user = await userModel.findOne({id: group.who_create_group});
+        await groupLeadersModel.create({id: buildToken(), user_name: user.name, group: group.name});
         return true;
     }
 
     async EditGroup(data: IGroupEditDTO): Promise<boolean> {
+        const userWhoEdit = await userModel.findOne({id: data.who_edit_group});
+        const group = await groupModel.findOne({id: data.id_group});
+
+        const userIsAuthorized = await groupLeadersModel.findOne({user_name: userWhoEdit.name, group: group.name});
+
+        if(!userIsAuthorized) {
+            throw new Error('Usuário não têm permissão para editar o grupo.');
+        }
+        
         const infoUpdated = await groupModel.findOneAndUpdate({id: data.id_group}, {
             name: data.name, 
             description: data.description,
@@ -60,6 +73,15 @@ export default class MongoRepository implements IUserRepository {
     }
 
     async RemoveGroup(data: IGroupRemoveDTO): Promise<boolean> {
+        const userWhoRemove = await userModel.findOne({id: data.who_remove_group});
+        const group = await groupModel.findOne({id: data.id_group});
+
+        const userIsAuthorized = await groupLeadersModel.findOne({user_name: userWhoRemove.name, group: group.name});
+
+        if(!userIsAuthorized) {
+            throw new Error('Usuário não têm permissão para remover o grupo.');
+        }
+
         const infoDeleted = await groupModel.deleteOne({id: data.id_group});
         
         if(infoDeleted.deletedCount === 0) {
@@ -75,6 +97,7 @@ export default class MongoRepository implements IUserRepository {
     }
 
     async removeEvent(data: IEventRemoveDTO): Promise<boolean> {
+        //ainda falta validar se o usuário têm permissão para remover
         const infoDeleted = await eventModel.deleteOne({id: data.event_id});
 
         if(infoDeleted.deletedCount === 0) {
@@ -85,6 +108,7 @@ export default class MongoRepository implements IUserRepository {
     }
 
     async editEvent(data: IEventEditDTO): Promise<boolean> {
+        //ainda falta validar se o usuário têm permissão para editar
         const infoUpdated = await eventModel.findOneAndUpdate({id: data.event_id}, {
             name: data.name,
             description: data.description,
@@ -95,16 +119,62 @@ export default class MongoRepository implements IUserRepository {
         if(infoUpdated.value === null) {
             throw new Error('Evento não encontrado.');
         }
-        //falta testar
+        
         return true;
     }
 
-    inviteToGroup(data: IUserInviteToGroupDTO): Promise<boolean> {
-        throw new Error("Method not implemented.");
+    async inviteToGroup(data: IUserInviteToGroupDTO): Promise<boolean> {
+        const userWhoInvite = await userModel.findOne({id: data.who_invite_id});
+        const group = await groupModel.findOne({name: data.group});
+        
+        const userIsAuthorized = await groupLeadersModel.findOne({user_name: userWhoInvite.name, group: group.name});
+
+        if(!userIsAuthorized) {
+            throw new Error('Usuário não têm permissão convidar para o grupo.');
+        }
+        const user = await userModel.findOne({id: data.who_invited_id});
+
+        if(!user) {
+            throw new Error('usuário não encontrado');
+        }
+
+        const userAlreadyInvitedToThisGroup = await groupSubscriberModel.findOne({
+            user_name: user.name,
+            group: data.group
+        });
+
+        if(userAlreadyInvitedToThisGroup) {
+            throw new Error('Este usuário já faz parte deste grupo.');
+        }
+
+        const groupExists = await this.findGroupByName(data.group);
+
+        if(!groupExists) {
+            throw new Error('Grupo informado não existe.');
+        }
+
+        await groupSubscriberModel.create({
+            id: buildToken(),
+            user_name: user.name,
+            group: data.group
+        });
+ 
+        return true;
     }
 
-    removeFromGroup(data: IUserRemoveFromGroupDTO): Promise<boolean> {
-        throw new Error("Method not implemented.");
+    async removeFromGroup(data: IUserRemoveFromGroupDTO): Promise<boolean> {
+        const userWhoInvite = await userModel.findOne({id: data.who_remove_id});
+        const group = await groupModel.findOne({name: data.group});
+        
+        const userIsAuthorized = await groupLeadersModel.findOne({user_name: userWhoInvite.name, group: group.name});
+        
+        if(!userIsAuthorized) {
+            throw new Error('Usuário não têm permissão para remover alguém do grupo.');
+        }
+
+        await groupSubscriberModel.deleteOne({id: data.who_removed_id});
+
+        return true;
     }
 
     async findGroupByName(name: string): Promise<boolean> {
